festival_2022/requirements.txt
asgiref==3.5.2
Django==4.1
django-carton==1.2.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
